+++ tools/slicec-cs/src/builders.rs
@@ -3,6 +3,7 @@
 use std::collections::HashMap;
 
 use crate::comments::{operation_parameter_doc_comment, CommentTag};
+use crate::cs_attributes;
 use crate::member_util::escape_parameter_name;
 use crate::slicec_ext::*;
 use slice::code_block::CodeBlock;
