+++ tools/slicec-cs/src/cs_attributes.rs
@@ -1,0 +1,9 @@
+// Copyright (c) ZeroC, Inc. All rights reserved.
+
+pub const ATTRIBUTE: &str = "cs::attribute";
+pub const ENCODED_RESULT: &str = "cs::encodedResult";
+pub const GENERIC: &str = "cs::generic";
+pub const INTERNAL: &str = "cs::internal";
+pub const NAMESPACE: &str = "cs::namespace";
+pub const READONLY: &str = "cs::readonly";
+pub const TYPE: &str = "cs::type";
