+++ tools/slicec-cs/src/slicec_ext/mod.rs
@@ -19,13 +19,3 @@ pub use operation_ext::OperationExt;
 pub use primitive_ext::PrimitiveExt;
 pub use slice_encoding_ext::EncodingExt;
 pub use type_ref_ext::TypeRefExt;
-
-pub mod cs_attributes {
-    pub const ATTRIBUTE: &str = "cs::attribute";
-    pub const ENCODED_RESULT: &str = "cs::encodedResult";
-    pub const GENERIC: &str = "cs::generic";
-    pub const INTERNAL: &str = "cs::internal";
-    pub const NAMESPACE: &str = "cs::namespace";
-    pub const READONLY: &str = "cs::readonly";
-    pub const TYPE: &str = "cs::type";
-}
